# perf(parser): Don't allocate for repr/decor for serde