--- a/crates/toml_edit/src/array.rs
+++ b/crates/toml_edit/src/array.rs
@@ -87,11 +87,6 @@ impl Array {
         &self.decor
     }
 
-    /// Returns the location within the original document
-    pub fn span(&self) -> Option<std::ops::Range<usize>> {
-        self.span.clone()
-    }
-
     pub(crate) fn despan(&mut self, input: &str) {
         self.span = None;
         self.decor.despan(input);