--- a/crates/toml_edit/src/document.rs
+++ b/crates/toml_edit/src/document.rs
@@ -51,8 +51,8 @@ impl Document {
     }
 
     /// Whitespace after last element
-    pub fn trailing(&self) -> &str {
-        self.trailing.as_str()
+    pub fn trailing(&self) -> &RawString {
+        &self.trailing
     }
 }