--- a/crates/toml_edit/src/inline_table.rs
+++ b/crates/toml_edit/src/inline_table.rs
@@ -169,8 +169,8 @@ impl InlineTable {
     }
 
     /// Whitespace after before element
-    pub fn preamble(&self) -> &str {
-        self.preamble.as_str()
+    pub fn preamble(&self) -> &RawString {
+        &self.preamble
     }
 
     /// Returns the location within the original document