--- a/crates/toml_edit/src/key.rs
+++ b/crates/toml_edit/src/key.rs
@@ -108,11 +108,6 @@ impl Key {
         &self.decor
     }
 
-    /// Returns the location within the original document
-    pub fn span(&self) -> Option<std::ops::Range<usize>> {
-        self.repr.as_ref().and_then(|r| r.span())
-    }
-
     pub(crate) fn despan(&mut self, input: &str) {
         self.decor.despan(input);
         if let Some(repr) = &mut self.repr {