--- a/crates/toml_edit/tests/testsuite/main.rs
+++ b/crates/toml_edit/tests/testsuite/main.rs
@@ -8,4 +8,5 @@ mod macros;
 mod parse;
 mod pretty;
 mod serde;
+mod span;
 mod stackoverflow;