… crates/toml_edit/src/parser/array.rs
… crates/toml_edit/src/parser/document.rs
… crates/toml_edit/src/parser/errors.rs
… crates/toml_edit/src/parser/inline_table.rs
… crates/toml_edit/src/parser/key.rs
… crates/toml_edit/src/parser/mod.rs
… crates/toml_edit/src/parser/state.rs
… crates/toml_edit/src/parser/table.rs
… crates/toml_edit/src/parser/trivia.rs
… crates/toml_edit/src/parser/value.rs
… crates/toml_edit/src/raw_string.rs
… crates/toml_edit/src/repr.rs
… crates/toml_edit/src/table.rs
… crates/toml_edit/src/value.rs
… crates/toml_edit/tests/fixtures/invalid/key/dotted-redefine-table.stderr
… crates/toml_edit/tests/testsuite/edit.rs
… crates/toml_edit/tests/testsuite/span.rs