Progress of performance through this PR
```
main:
de/minimal: 3.23 us
Value/minimal: 3.73 us
de/medium: 131.97 us
Value/medium: 129.48 us

d32e72a: refactor(parser): Track parse location
de/minimal: 3.98 us
Value/minimal: 4.00 us
de/medium: 142.97 us
Value/medium: 150.83 us

7a40a0770: feat(parser): Track span of changes
de/minimal: 3.82 us
Value/minimal: 4.26 us
de/medium: 141.61 us
Value/medium: 119.87 us

7ec8d71: feat(parser): Track spans for every field
de/minimal: 4.05 us
Value/minimal: 4.57 us
de/medium: 149.47 us
Value/medium: 133.46 us

4c9c71b: refactor(parser): Only track spans
de/minimal: 4.11 us
Value/minimal: 4.29 us
de/medium: 129.30 us
Value/medium: 128.25 us

3591592e8: perf(easy): Leave spans alone
de/minimal: 3.36 us
Value/minimal: 4.05 us
de/medium: 105.85 us
Value/medium: 104.95 us
```

I looked at the serde performance as that is the one we care most about.  When people are roundtripping or looking at raw data, then generally performance doesn't matter as much.  Or rather, this is the main case cargo would care about.

Switching the parsers input to `Located` slowed things down.  We made up for that in `medium` but never did in `minimal`.  It would be nice to drop that overhead.  nom-experimental#61 explores this further.

I also tried speeding up the `edit` workflow by using `str::get_unchecked` but it didn't offer a noticeable improvement.